move sendrequest into client
