+++ cmd/habitatctl/commands/community.go
@@ -20,6 +20,8 @@ import (
 
 	"github.com/eagraf/habitat/structs/ctl"
 	"github.com/spf13/cobra"
+
+	client "github.com/eagraf/habitat/pkg/habitat_client"
 )
 
 // communityCmd represents the community command
@@ -50,10 +52,10 @@ var communityCmd = &cobra.Command{
 				if len(args) < 3 {
 					fmt.Printf("No member id specified for community add command")
 				}
-				SendRequest(ctl.CommandCommunityAddMember, []string{args[1], args[2]})
+				client.SendRequest(ctl.CommandCommunityAddMember, []string{args[1], args[2]})
 				return
 			case "propose":
-				SendRequest(ctl.CommandCommunityPropose, []string{})
+				client.SendRequest(ctl.CommandCommunityPropose, []string{})
 				return
 			default:
 				fmt.Printf("%s is an invalid subcommand for community %s\n", args[1], args[0])
@@ -75,7 +77,7 @@ var communityCreateCmd = &cobra.Command{
 			return
 		}
 
-		SendRequest(ctl.CommandCommunityCreate, []string{name.Value.String(), address.Value.String()})
+		client.SendRequest(ctl.CommandCommunityCreate, []string{name.Value.String(), address.Value.String()})
 	},
 }
 
@@ -96,7 +98,7 @@ var communityJoinCmd = &cobra.Command{
 			return
 		}
 
-		SendRequest(ctl.CommandCommunityJoin, []string{address.Value.String(), communityID.Value.String()})
+		client.SendRequest(ctl.CommandCommunityJoin, []string{address.Value.String(), communityID.Value.String()})
 	},
 }
 
@@ -122,7 +124,7 @@ var communityAddMemberCmd = &cobra.Command{
 			return
 		}
 
-		SendRequest(ctl.CommandCommunityAddMember, []string{
+		client.SendRequest(ctl.CommandCommunityAddMember, []string{
 			communityID.Value.String(),
 			nodeID.Value.String(),
 			address.Value.String(),
@@ -146,7 +148,7 @@ var communityProposeTransitionCmd = &cobra.Command{
 		}
 		b64Patch := args[0]
 
-		SendRequest(ctl.CommandCommunityPropose, []string{communityID.Value.String(), b64Patch})
+		client.SendRequest(ctl.CommandCommunityPropose, []string{communityID.Value.String(), b64Patch})
 	},
 }
 
@@ -160,7 +162,7 @@ var communityStateCmd = &cobra.Command{
 			return
 		}
 
-		SendRequest(ctl.CommandCommunityState, []string{communityID.Value.String()})
+		client.SendRequest(ctl.CommandCommunityState, []string{communityID.Value.String()})
 	},
 }
 
@@ -168,7 +170,7 @@ var communityListCmd = &cobra.Command{
 	Use:   "ls",
 	Short: "list the communities that this node is a part of",
 	Run: func(cmd *cobra.Command, args []string) {
-		SendRequest(ctl.CommandCommunityList, []string{})
+		client.SendRequest(ctl.CommandCommunityList, []string{})
 	},
 }
