+++ cmd/habitatctl/commands/ps.go
@@ -16,6 +16,7 @@ limitations under the License.
 package commands
 
 import (
+	client "github.com/eagraf/habitat/pkg/habitat_client"
 	"github.com/spf13/cobra"
 )
 
@@ -25,7 +26,7 @@ var psCmd = &cobra.Command{
 	Short: "List running habitat processes",
 
 	Run: func(cmd *cobra.Command, args []string) {
-		SendRequest("ps", args)
+		client.SendRequest("ps", args)
 	},
 }
