+++ cmd/habitatctl/commands/stop.go
@@ -1,6 +1,7 @@
 package commands
 
 import (
+	client "github.com/eagraf/habitat/pkg/habitat_client"
 	"github.com/spf13/cobra"
 )
 
@@ -11,7 +12,7 @@ var stopCmd = &cobra.Command{
 	Long:  `TODO create long description`,
 	Args:  cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
-		SendRequest("stop", args)
+		client.SendRequest("stop", args)
 	},
 }
