+++ pkg/habitat_client/client.go
@@ -10,16 +10,17 @@ import (
 	"github.com/eagraf/habitat/structs/ctl"
 )
 
+// TODO: don't hard code this value in
 const (
-	ClientHost = "localhost"
+	ClientHost = "localhost:2040"
 )
 
 type Client struct {
 	conn net.Conn
 }
 
-func NewClient(port string) (*Client, error) {
-	conn, err := net.Dial("tcp", fmt.Sprintf("%s:%s", ClientHost, port))
+func NewClient() (*Client, error) {
+	conn, err := net.Dial("tcp", ClientHost)
 	if err != nil {
 		return nil, err
 	}
@@ -68,4 +69,26 @@ func (c *Client) ReadResponse() (*ctl.Response, error) {
 	c.conn.Close()
 
 	return &res, nil
+}
+
+func SendRequest(command string, args []string) (*ctl.Response, error) {
+	client, err := NewClient()
+	if err != nil {
+		fmt.Println("Error: couldn't connect to habitat service")
+		return nil, err
+	}
+
+	err = client.WriteRequest(&ctl.Request{
+		Command: command,
+		Args:    args,
+	})
+	if err != nil {
+		fmt.Printf("Error creating request to habitat service: %s", err)
+	}
+
+	res, err := client.ReadResponse()
+	if err != nil {
+		fmt.Printf("Error: couldn't read response from habitat service: %s\n", err)
+	}
+	return res, err
 }
